Created LambdaRekognition folder and updated Lambda Rekognition script file
